adjustment: small repositioning of elements
